pyproject.toml
[tool.poetry]
name = "hisobchi-ai"
version = "0.1.0"
description = "Hisobchi.ai - Telegram bot and REST API for personal finance tracking"
authors = ["Sukhrob <sukhrob@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiogram = "^3.3.0"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.6"
pydantic = {version = "<2.12", extras = ["email"]}
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.6.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
